requirements.txt
langchain==0.1.12
langchain-google-genai==0.0.11
google-generativeai>=0.4.1,<0.5.0
python-dotenv==1.0.1
PyPDF2>=3.0.0
